+++ R/record_status.R
@@ -1,34 +1,34 @@
-#' @title Record package issues.
+#' @title Record package status.
 #' @export
 #' @keywords package check data management
-#' @description Record R-multiverse package issues in
+#' @description Record R-multiverse package status in
 #'   package-specific JSON files.
-#' @section Package issues:
-#'   Functions like [issues_versions()] and [issues_descriptions()]
+#' @section Package status:
+#'   Functions like [status_versions()] and [status_descriptions()]
 #'   perform health checks for all packages in R-multiverse.
 #'   For a complete list of checks, see
-#'   the `issues_*()` functions listed at
+#'   the `status_*()` functions listed at
 #'   <https://r-multiverse.org/multiverse.internals/reference/index.html>.
 #'   [record_versions()] updates the version number history
-#'   of releases in R-multiverse, and [record_issues()] gathers
-#'   together all the issues about R-multiverse packages.
-#' @section Issue data:
-#'   For each package with observed problems, [record_issues()] writes
+#'   of releases in R-multiverse, and [record_status()] gathers
+#'   together all the status about R-multiverse packages.
+#' @section Status data:
+#'   For each package with observed problems, [record_status()] writes
 #'   a JSON list entry in the output JSON file
 #'   with one element for each type of failing check.
 #'   Each check-specific element has an informative name
 #'   (for example, `checks`, `descriptions`, or `versions`)
 #'   and a list of diagnostic information. In addition, there is a `date`
 #'   field to indicate when an issue was first detected. The `date`
-#'   automatically resets the next time all the issues in the package
+#'   automatically resets the next time all the status in the package
 #'   are resolved.
 #' @return `NULL` (invisibly).
-#' @inheritParams issues_checks
-#' @inheritParams issues_dependencies
-#' @inheritParams issues_versions
+#' @inheritParams status_checks
+#' @inheritParams status_dependencies
+#' @inheritParams status_versions
 #' @inheritParams meta_checks
 #' @param output Character of length 1, file path to the JSON file to record
-#'   new package issues. Each call to `record_issues()` overwrites the
+#'   new package status. Each call to `record_status()` overwrites the
 #'   contents of the file.
 #' @param mock For testing purposes only, a named list of data frames
 #'   for inputs to various intermediate functions.
@@ -40,40 +40,40 @@
 #'     versions = versions,
 #'     repo = repo
 #'   )
-#'   record_issues(
+#'   record_status(
 #'     repo = repo,
 #'     versions = versions,
 #'     output = output
 #'   )
 #'   writeLines(readLines(output))
-record_issues <- function(
+record_status <- function(
   repo = "https://community.r-multiverse.org",
   versions = "versions.json",
-  output = "issues.json",
+  output = "status.json",
   mock = NULL,
   verbose = FALSE
 ) {
   today <- mock$today %||% format(Sys.Date(), fmt = "yyyy-mm-dd")
   checks <- mock$checks %||% meta_checks(repo = repo)
   packages <- mock$packages %||% meta_packages(repo = repo)
-  issues <- list() |>
-    add_issues(issues_checks(meta = checks), "checks") |>
-    add_issues(issues_descriptions(meta = packages), "descriptions") |>
-    add_issues(issues_versions(versions = versions), "versions")
-  issues <- issues |>
-    add_issues(
-      issues_dependencies(names(issues), packages, verbose = verbose),
+  status <- list() |>
+    add_status(status_checks(meta = checks), "checks") |>
+    add_status(status_descriptions(meta = packages), "descriptions") |>
+    add_status(status_versions(versions = versions), "versions")
+  status <- status |>
+    add_status(
+      status_dependencies(names(status), packages, verbose = verbose),
       "dependencies"
     )
-  for (name in names(issues)) {
-    issues[[name]]$success <- FALSE
+  for (name in names(status)) {
+    status[[name]]$success <- FALSE
   }
-  for (healthy in setdiff(packages$package, names(issues))) {
-    issues[[healthy]] <- list(success = TRUE)
+  for (healthy in setdiff(packages$package, names(status))) {
+    status[[healthy]] <- list(success = TRUE)
   }
-  issues <- issues[sort(names(issues))]
-  overwrite_issues(
-    issues = issues,
+  status <- status[sort(names(status))]
+  overwrite_status(
+    status = status,
     output = output,
     today = today,
     packages = packages
@@ -81,23 +81,23 @@ record_issues <- function(
   invisible()
 }
 
-add_issues <- function(total, subset, category) {
+add_status <- function(total, subset, category) {
   for (package in names(subset)) {
     total[[package]][[category]] <- subset[[package]]
   }
   total
 }
 
-overwrite_issues <- function(issues, output, today, packages) {
+overwrite_status <- function(status, output, today, packages) {
   previous <- list()
   if (file.exists(output)) {
     previous <- jsonlite::read_json(output, simplifyVector = TRUE)
   }
-  for (package in names(issues)) {
-    issues[[package]]$date <- previous[[package]]$date %||% today
+  for (package in names(status)) {
+    status[[package]]$date <- previous[[package]]$date %||% today
     index <- packages$package == package
-    issues[[package]]$version <- packages$version[index]
-    issues[[package]]$remote_hash <- packages$remotesha[index]
+    status[[package]]$version <- packages$version[index]
+    status[[package]]$remote_hash <- packages$remotesha[index]
   }
-  jsonlite::write_json(x = issues, path = output, pretty = TRUE)
+  jsonlite::write_json(x = status, path = output, pretty = TRUE)
 }